requirements.txt
# DjCap - djay Pro Capture and Analysis
# Dependencies for capturing djay Pro window and extracting metadata

# Window capture
mss>=9.0.1
Pillow>=10.0.0

# macOS Core Graphics (optional, for better window capture)
pyobjc-framework-Quartz>=10.0

# OCR using Apple Vision
ocrmac>=0.1.0

# Image processing
numpy>=1.24.0
opencv-python>=4.8.0

